+++ src/runtime.ts
@@ -119,4 +119,5 @@ export { config } from './config';
 export * from './client';
 export * from './i18n';
 export * from './media';
+export * from './preview';
 export type * from './types';